backend/requirements.txt
fastapi
uvicorn
web3
zama-concrete
